requirements.txt
aiogram==3.22.0
python-environ==0.4.54
requests==2.32.5